+++ azure-pipelines.yml
@@ -1,7 +1,7 @@
 # https://aka.ms/yaml
-trigger:
-- master
-- releases/*
+# trigger:
+# - master
+# - releases/*
 
 pool:
   vmImage: 'Ubuntu 16.04'
